Bitcoin uses Base58Address
